+++ src/Products/MeetingCommunes/tests/testCustomViews.py
@@ -677,7 +677,7 @@ class testCustomViews(MeetingCommunesTestCase):
     def test__filter_items(self):
         cfg = self.meetingConfig
         self.changeUser('pmManager')
-        cfg.setUseGroupsAsCategories(False)
+        self._enableField('category')
         cfg.setInsertingMethodsOnAddItem(
             ({'insertingMethod': 'on_categories', 'reverse': '0'},))
         m = self._createMeetingWithItems()
@@ -846,7 +846,7 @@ class testCustomViews(MeetingCommunesTestCase):
     def test_get_grouped_items(self):
         self.changeUser('pmManager')
         cfg = self.meetingConfig
-        cfg.setUseGroupsAsCategories(False)
+        self._enableField('category')
         cfg.setInsertingMethodsOnAddItem(
             ({'insertingMethod': 'on_categories', 'reverse': '0'},))
         m = self._createMeetingWithItems()
@@ -933,6 +933,42 @@ class testCustomViews(MeetingCommunesTestCase):
              [u'Developers', [u'Development topics', [items[3]]]],
              [u'Vendors', [u'Development topics', [items[4]]]],
              [u'Developers', [developpers.Title(), [i5]], [u'Research topics', [i7]]]])
+
+        # using _group_by_ function instead of persistent values on items
+        self.changeUser('siteadmin')
+        new_ss_org = self.create('organization', folder=self.developers, id='sous-org', title='Sous Org', acronym='SORG')
+        new_ss_org_uid = new_ss_org.UID()
+        self._select_organization(new_ss_org_uid)
+        items[3].setProposingGroup(new_ss_org_uid)
+        items[3]._update_after_edit()
+        self.changeUser('pmManager')
+        res = helper.get_grouped_items(itemUids, group_by=['org_first_level_title'])
+        self.assertListEqual(
+            res,
+            [['Developers', [items[0]]],
+             ['Vendors', items[1:3]],
+             ['Developers', [items[3]]],
+             ['Vendors', [items[4]]],
+             ['Developers', items[5:7]]])
+
+        res = helper.get_grouped_items(itemUids, group_by=['org_first_level'])
+        # the proposing group for item3 is "Sous org" but we group by org_first_level --> it's the "Developers" group
+        self.assertListEqual(
+            res,
+            [[items[0].getProposingGroup(theObject=True), [items[0]]],
+             [items[1].getProposingGroup(theObject=True), items[1:3]],
+             [items[0].getProposingGroup(theObject=True), [items[3]]],
+             [items[4].getProposingGroup(theObject=True), [items[4]]],
+             [items[5].getProposingGroup(theObject=True), items[5:7]]])
+
+        res = helper.get_grouped_items(itemUids, group_by=['proposingGroup'])
+        self.assertListEqual(
+            res,
+            [[u'Developers', [items[0]]],
+             [u'Vendors', items[1:3]],
+             [u'Developers / Sous Org', [items[3]]],
+             [u'Vendors', [items[4]]],
+             [u'Developers', items[5:7]]])
 
     def test_get_grouped_items_unrestricted(self):
         self.changeUser('pmManager')
@@ -1040,7 +1076,7 @@ class testCustomViews(MeetingCommunesTestCase):
     def test_get_multiple_level_printing(self):
         self.changeUser('pmManager')
         cfg = self.meetingConfig
-        cfg.setUseGroupsAsCategories(False)
+        self._enableField('category')
         cfg.setInsertingMethodsOnAddItem(
             ({'insertingMethod': 'on_categories', 'reverse': '0'},))
         m = self._createMeetingWithItems()
@@ -1094,7 +1130,7 @@ class testCustomViews(MeetingCommunesTestCase):
 
     def test_print_item_number_within_category(self):
         cfg = self.meetingConfig
-        cfg.setUseGroupsAsCategories(True)
+        self._enableField('category', enable=False)
 
         def create_and_validate_item(creator, preferred_meeting=None):
             self.changeUser(creator)